pyproject.toml
[project]
name = "webu"
version = "0.8.0"
authors = [
    { name="Hansimov" },
]
description = "Web Utils for browsing and scraping"
readme = "README.md"
license = "MIT"
requires-python = ">=3.6"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [ "tclogger", "DrissionPage", "pyvirtualdisplay", "beautifulsoup4", "requests", "netifaces", "fastapi", "uvicorn", "numpy", "playwright", "psutil", "pymongo", "pytest-asyncio", "aiohttp", "aiohttp-socks", "undetected-chromedriver", "selenium"]

[project.urls]
Homepage = "https://github.com/Hansimov/webu"
Issues = "https://github.com/Hansimov/webu/issues"

[project.scripts]
ggsc = "webu.google_api.cli:main"
pxsc = "webu.proxy_api.cli:main"
